pyproject.toml
[project]
name = "gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-api-python-client>=2.177.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.2",
    "mcp>=1.12.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
